remove unused TableCell method, move method from ExcelTable to AbstractTable
